+++ R/package-docs.R
@@ -134,8 +134,9 @@
 #'
 #'   In some rmarkdown and LaTeX formats, you also need to add LaTeX dependencies
 #'   manually. Run [report_latex_dependencies()] and add
-#'   the output to your LaTeX preamble, or in Rmarkdown formats, add it to the
-#'   rmarkdown header like this:
+#'   the output to your LaTeX preamble. Also add the command definitions printed
+#'   by `cat(latex_commands(), sep = "\\n")`. In Rmarkdown formats, add
+#'   the package dependencies to the rmarkdown header like this:
 #'
 #'   ```
 #'   header-includes: